src/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["avenir_web*"]
exclude = ["config*"]

[project]
name = "avenir-web"
version = "0.1.0"
description = "Avenir-Web is a Python web-agent runner that uses Playwright and large multimodal models (via LiteLLM) to autonomously carry out tasks on live websites."
authors = [
    {name = "Aiden Yiliu Li", email = "yiliu.li@outlook.com"},
    {name = "Xinyue Hao", email = "x.hao-5@sms.ed.ac.uk"},
    {name = "Shilong Liu", email = "slongliu86@gmail.com"},
    {name = "Mengdi Wang", email = "mengdiw@princeton.edu"}
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "httpx==0.27.2",
    "litellm==1.35.32",
    "backoff==2.2.1",
    "python-dotenv==1.0.1",
    "playwright>=1.50.0",
    "jsonlines>=3.1.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "protobuf>=4.21.0",
    "beautifulsoup4>=4.12.0",
    "lxml==4.9.3",
    "toml>=0.10.2",
    "aioconsole>=0.6.0",
    "psutil>=5.9.0",
    "inquirerpy==0.3.4",
    "pillow>=10.0.0",
    "requests>=2.31.0"
]
